% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationverse-package.R
\docType{package}
\name{mediationverse-package}
\alias{mediationverse}
\alias{mediationverse-package}
\title{mediationverse: Ecosystem for Mediation Analysis in R}
\description{
The mediationverse is a collection of R packages for mediation analysis.
This meta-package loads all core packages with a single \code{library()} command.
}
\details{
\subsection{Core Packages}{

The mediationverse includes the following packages:
\itemize{
\item \strong{medfit}: Infrastructure for model fitting, extraction, and bootstrap inference
\item \strong{probmed}: Probabilistic effect size (P_med)
\item \strong{RMediation}: Confidence intervals (DOP, MBCO, MC methods)
\item \strong{medrobust}: Sensitivity analysis for unmeasured confounding
\item \strong{medsim}: Simulation infrastructure for mediation research
\item \strong{missingmed}: Mediation analysis with multiple imputation and IPW for missing data
}
}

\subsection{Installation}{

Install all packages at once:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{install.packages("mediationverse")
}\if{html}{\out{</div>}}

Or install development versions from GitHub:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pak::pak("Data-Wise/mediationverse")
}\if{html}{\out{</div>}}
}

\subsection{Usage}{

Load all packages:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(mediationverse)
}\if{html}{\out{</div>}}

Check installed packages:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mediationverse_packages()
}\if{html}{\out{</div>}}

Update all packages:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mediationverse_update()
}\if{html}{\out{</div>}}
}

\subsection{Learn More}{
\itemize{
\item medfit: \url{https://data-wise.github.io/medfit/}
\item probmed: \url{https://data-wise.github.io/probmed/}
\item RMediation: \url{https://cran.r-project.org/package=RMediation}
\item medrobust: \url{https://data-wise.github.io/medrobust/}
\item medsim: \url{https://data-wise.github.io/medsim/}
\item missingmed: \url{https://data-wise.github.io/missingmed/}
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://Data-Wise.github.io/mediationverse/}
  \item \url{https://github.com/Data-Wise/mediationverse}
  \item Report bugs at \url{https://github.com/Data-Wise/mediationverse/issues}
}

}
\author{
\strong{Maintainer}: Davood Tofighi \email{dtofighi@gmail.com} (\href{https://orcid.org/0000-0001-8523-7776}{ORCID})

Authors:
\itemize{
  \item Davood Tofighi \email{dtofighi@gmail.com} (\href{https://orcid.org/0000-0001-8523-7776}{ORCID})
}

}
\keyword{internal}
