% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationverse-package.R
\docType{package}
\name{mediationverse-package}
\alias{mediationverse-package}
\title{mediationverse: Ecosystem for Mediation Analysis in R}
\description{
A meta-package that loads the core packages in the mediation analysis ecosystem: 'medfit' (infrastructure), 'probmed' (probabilistic effect sizes), 'RMediation' (confidence intervals), and 'medrobust' (sensitivity analysis). This package simplifies installation and provides a unified entry point for mediation analysis in R. Inspired by the 'tidyverse' and 'easystats' ecosystems.

The mediationverse is a collection of R packages for mediation analysis.
This meta-package loads the core packages with a single command.
}
\details{
\subsection{Core Packages}{

The mediationverse includes the following packages:
\itemize{
\item \strong{medfit}: Infrastructure for model fitting and extraction
\item \strong{probmed}: Probabilistic effect size (P_med)
\item \strong{RMediation}: Confidence intervals (DOP, MBCO, MC methods)
\item \strong{medrobust}: Sensitivity analysis for unmeasured confounding
}
}

\subsection{Installation}{

Install all packages at once:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{install.packages("mediationverse")
}\if{html}{\out{</div>}}
}

\subsection{Usage}{

Load all packages:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(mediationverse)
}\if{html}{\out{</div>}}
}

\subsection{Learn More}{
\itemize{
\item medfit: \url{https://data-wise.github.io/medfit/}
\item probmed: \url{https://data-wise.github.io/probmed/}
\item RMediation: \url{https://cran.r-project.org/package=RMediation}
\item medrobust: \url{https://data-wise.github.io/medrobust/}
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data-wise/mediationverse}
  \item \url{https://data-wise.github.io/mediationverse/}
  \item Report bugs at \url{https://github.com/data-wise/mediationverse/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/data-wise/mediationverse}
  \item \url{https://data-wise.github.io/mediationverse/}
  \item Report bugs at \url{https://github.com/data-wise/mediationverse/issues}
}

}
\author{
\strong{Maintainer}: Davood Tofighi \email{dtofighi@gmail.com} (\href{https://orcid.org/0000-0001-8523-7776}{ORCID})

}
\keyword{internal}
